% Generated by roxygen2 (3.2.0): do not edit by hand
% Please edit documentation in R/generer_grille_picross.R


\name{generer_grille_picross}
\alias{generer_grille_picross}
\title{Générer une grille de Picross aléatoire}
\usage{
generer_grille_picross(taille)
}
\arguments{
  \item{taille}{La taille de la grille}
}
\description{
Cette fonction génère une grille de Picross aléatoire de la taille spécifiée.
}
\examples{
generer_grille_picross(5)
}



% Generated by roxygen2 (3.2.0): do not edit by hand
% Please edit documentation in R/generer_indices.R

\name{generer_indices}
\alias{generer_indices}
\title{Générer des indices aléatoires pour les lignes et les colonnes}
\usage{
generer_indices(taille)
}
\arguments{
  \item{taille}{La taille de la grille}
}
\description{
Cette fonction génère des indices aléatoires pour les lignes et les colonnes d'une grille de Picross.
}
\examples{
generer_indices(5)
}


% Generated by roxygen2 (3.2.0): do not edit by hand
% Please edit documentation in R/verifier_solution.R


\name{verifier_solution}
\alias{verifier_solution}
\title{Vérifier si la solution de la grille est correcte}
\usage{
verifier_solution(grille, indices_lignes, indices_colonnes)
}
\arguments{
  \item{grille}{La grille de Picross remplie}
  \item{indices_lignes}{Les indices pour les lignes de la grille}
  \item{indices_colonnes}{Les indices pour les colonnes de la grille}
}
\description{
Cette fonction vérifie si la solution proposée pour la grille de Picross est correcte en comparant les indices fournis avec la grille remplie.
}
\examples{
grille <- generer_grille_picross(5)
indices <- generer_indices(5)
verifier_solution(grille, indices, indices)
}


% Generated by roxygen2 (3.2.0): do not edit by hand
% Please edit documentation in R/App.R


\name{ui}
\alias{ui}
\title{Interface utilisateur pour le jeu de Picross}
\description{
Cette fonction définit l'interface utilisateur (UI) pour le jeu de Picross en utilisant Shiny.
}
\usage{
ui
}
\examples{
ui
}



% Generated by roxygen2 (3.2.0): do not edit by hand
% Please edit documentation in R/App.R


\name{server}
\alias{server}
\title{Logique du serveur pour le jeu de Picross}
\description{
Cette fonction définit la logique du serveur pour le jeu de Picross en utilisant Shiny. Elle réagit aux actions de l'utilisateur et met à jour l'état de la grille de jeu.
}
\usage{
server(input, output, session)
}
\arguments{
  \item{input}{Les données d'entrée de l'application Shiny}
  \item{output}{Les résultats de l'application Shiny à envoyer à l'interface utilisateur}
  \item{session}{La session Shiny en cours}
}
\examples{
server(input, output, session)
}
